tests: split tests and refactor with SdkBaseTest
