+++ dash-sdk-java/src/main/swig/root.i
@@ -41,6 +41,7 @@ MemoryFactory & memoryFactory = *MemoryFactory::getInstance();
 %include "generics/byte_array.i"
 
 // Platform Related Structures
+%include "sdk.i"
 %include "i128.i"
 %include "binary_data.i"
 %include "contract_bounds.i"
@@ -86,6 +87,8 @@ import org.dashj.platform.sdk.base.BaseObject;
 %}
 // ignore is generated by ignore.py
 %include "ignore.i"
+
+%include "tests.i"
 
 // header files to convert to Java
 %include "../../../../dash-sdk-bindings/target/dash_sdk_bindings.h"
