+++ dash-sdk-java/src/main/swig/tests.i
@@ -1,0 +1,4 @@
+
+%newobject platform_mobile_tests_get_binary_data_empty();
+%newobject platform_mobile_tests_get_platform_value_bool(bool value);
+%newobject platform_mobile_tests_get_binary_data_4();
