+++ dash-sdk-java/src/test/java/org/dashj/platform/sdk/BinaryDataTest.java
@@ -1,11 +1,11 @@
 package org.dashj.platform.sdk;
 
-import org.junit.jupiter.api.Test;
+import org.junit.Test;
 
-import static org.junit.jupiter.api.Assertions.assertArrayEquals;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 
 public class BinaryDataTest extends BaseTest {
     @Test
@@ -14,7 +14,7 @@ public class BinaryDataTest extends BaseTest {
         assertTrue(data.swigCMemOwn);
         assertEquals(32, data.get_0().length);
         assertArrayEquals(identifier, data.get_0());
-        data.delete(); // does not crash
+        data.delete();
     }
 
     // if the object is created in Java, then the Java object does own it
@@ -26,6 +26,24 @@ public class BinaryDataTest extends BaseTest {
         assertEquals(32, data.get_0().length);
         assertArrayEquals(identifier, data.get_0());
         System.out.printf("BinaryData %x\n", BinaryData.getCPtr(data));
-        data.delete(); // does not crash
+        data.delete();
+    }
+
+    @Test
+    public void getBinaryDataEmptyTest() {
+        BinaryData data = dashsdk.platformMobileTestsGetBinaryDataEmpty();
+        assertNotNull(data);
+        assertEquals(0, data.get_0().length);
+        assertArrayEquals(new byte[0], data.get_0());
+        data.delete();
+    }
+
+    @Test
+    public void getBinaryData4Test() {
+        BinaryData data = dashsdk.platformMobileTestsGetBinaryData4();
+        assertNotNull(data);
+        assertEquals(4, data.get_0().length);
+        assertArrayEquals(new byte[] {0, 1, 2, 3}, data.get_0());
+        data.delete();
     }
 }
