+++ dash-sdk-java/src/test/java/org/dashj/platform/sdk/CoreTest.java
@@ -1,27 +1,17 @@
 package org.dashj.platform.sdk;
 
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
+
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.Test;
 
 import java.math.BigInteger;
 
-import static org.junit.jupiter.api.Assertions.assertArrayEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertNotNull;
 
-public class CoreTest extends BaseTest {
 
-    static SWIGTYPE_p_DashSdk sdk;
-
-    @BeforeAll
-    static void startUp() {
-        sdk = dashsdk.platformMobileSdkCreateDashSdk(BigInteger.ZERO, BigInteger.ZERO, true);
-    }
-
-    @AfterAll
-    static void tearDown() {
-        dashsdk.platformMobileSdkDestroyDashSdk(sdk);
-    }
+public class CoreTest extends SdkBaseTest {
 
     public static byte[] hexStringToByteArray(String hex) {
         int len = hex.length();
@@ -34,7 +24,7 @@ public class CoreTest extends BaseTest {
     }
 
     @Test
-    void getTransactionTest() {
+    public void getTransactionTest() {
         byte[] txid = hexStringToByteArray("13155120729d7ee473e4eb8c71abd5a70370cade586b0c34120c5e0e2c3f0e48");
         Result_ok_Vec_u8_err_String result = dashsdk.platformMobileCoreGetTransactionSdk(sdk, txid);
         assertNotNull(result);
