+++ dash-sdk-java/src/test/java/org/dashj/platform/sdk/IdentifierTest.java
@@ -1,11 +1,9 @@
 package org.dashj.platform.sdk;
 
-import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertArrayEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import org.junit.Test;
+
+import static org.junit.Assert.assertArrayEquals;
 
 public class IdentifierTest extends BaseTest {
     @Test
