+++ platform-mobile/src/lib.rs
@@ -13,35 +13,7 @@ pub mod core;
 mod logs;
 pub mod voting;
 pub mod sdk;
+pub mod tests;
 
 extern crate ferment_macro;
 
-use drive_proof_verifier::ContextProvider;
-use platform_value::types::binary_data::BinaryData;
-use dash_sdk::platform::types::identity::PublicKeyHash;
-use dpp::bincode::{Decode, Encode};
-use platform_value::{Hash256, Value};
-
-#[ferment_macro::export]
-pub fn convert_to_pkh(pkh: [u8; 20]) -> PublicKeyHash {
-    PublicKeyHash(pkh)
-}
-
-#[ferment_macro::export]
-pub fn get_binary_data() -> BinaryData {
-    BinaryData::new(vec![])
-}
-#[ferment_macro::export]
-pub fn get_binary_data2() -> BinaryData {
-     BinaryData(vec![0, 1, 2, 3])
-}
-
-#[ferment_macro::export]
-pub fn get_platform_value() -> Value {
-    Value::Bool(false)
-}
-
-#[ferment_macro::export]
-pub fn get_platform_value_with_map() ->  Value {
-    Value::Map(vec![(Value::Text("key".to_string()), Value::I32(4))])
-}
